apps/ai-manager/pyproject.toml
[project]
name = "ai-manager"
version = "0.1.0"
description = "AI manager server used on Hyperbolic for AI inference"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "pytest-cov>=4.1.0",
]
